requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.0.15
chromadb>=0.4.22
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.0
langsmith>=0.0.69
